+++ src/main.ts
@@ -1,3 +1,5 @@
+import StringUtilsWord from './word';
+
 export class StringUtils {
   /**
    * Check if a given string is blank or not
@@ -7,6 +9,73 @@ export class StringUtils {
    */
   public static isBlank(str: string): boolean {
     return str.trim().replaceAll(' ', '').length == 0;
+  }
+
+  /**
+   * Returns a boolean indicating if given string is atleast 2 lengthed string
+   *
+   * @param {string} str
+   * @example
+   * str: '      t'
+   * returns: false
+   * @example
+   * str: 'test'
+   * returns: true
+   * @example
+   * str: 'of    '
+   * returns: true
+   */
+  public static isConsiderableCharSequence(str: string): boolean {
+    return str.trim().replaceAll(' ', '').length >= 2;
+  }
+
+  /**
+   * Returns a boolean indicating if a given string table contains atleast one considerable subsequence
+   *
+   * @param {string[]} stringTable - Should be a table of string
+   *
+   * @example
+   * stringTable: ['This', 'is', 'my', 'test']
+   * retunrs: true
+   * @example
+   * stringTable: [' ', ' ', ' ']
+   * retunrs: false
+   * @example
+   * stringTable: ['t', 'h', 'i']
+   * retruns: false
+   */
+  public static containsConsiderableCharSequence(
+    stringTable: string[],
+  ): boolean {
+    if (this.isBlank(stringTable.join(''))) return false;
+    return (
+      stringTable.find((subsequence) =>
+        this.isConsiderableCharSequence(subsequence),
+      ) != undefined
+    );
+  }
+
+  /**
+   * Returns a boolean indicating if a given string table contains only considerable subsequence
+   *
+   * @param {string[]} stringTable - Should be a table of string
+   *
+   * @example
+   * stringTable: ['This', 'is', 'my', 'example']
+   * returns: true
+   * @example
+   * stringTable: ['This', 'i', 's', 'my', 'example']
+   * returns: false
+   */
+  public static containsOnlyConsiderableCharSequences(
+    stringTable: string[],
+  ): boolean {
+    if (this.isBlank(stringTable.join(''))) return false;
+    return (
+      stringTable.find(
+        (subsequence) => !this.isConsiderableCharSequence(subsequence),
+      ) == undefined
+    );
   }
 
   /**
@@ -33,5 +102,64 @@ export class StringUtils {
     strArray[index] = toReplaceWith;
 
     return strArray.join('');
+  }
+
+  /**
+   * Return a string removed from all blank chars
+   *
+   * @param {string} str - The string of which remove blank spaces
+   *
+   * @example
+   * str: This is my example
+   * returns: Thisismyexample
+   */
+  public static removeBlankChars(str: string): string {
+    return str.replaceAll(' ', '');
+  }
+
+  /**
+   * Returns a table of string where each element is at least 2 length char
+   *
+   * @param {string} str - Should be a string containing spaces and at least 2 letters
+   *
+   * @remarks
+   * This method works only if str starts with a relevant / considerable sub string sequence, it should be rework later to manage following relevant sequences
+   *
+   * @example
+   * str: 'This    is my   ex a m p l   e'
+   * returns: ['This', 'is', 'my', 'example']
+   * @example
+   * str: 'T h i s m y e x a m p l e'
+   * returns: ['Thisismyexample']
+   * @example
+   * str: 'Thi   s is my exa mple'
+   * returns: ['This is my exa mple']
+   *
+   */
+  public static blendIrrelevantStringsInRelevantOnes(str: string): string[] {
+    const splittedStr =
+      StringUtilsWord.normalizeSpacesBetweenWords(str).split(' ');
+
+    if (!this.containsConsiderableCharSequence(splittedStr))
+      return [this.removeBlankChars(str)];
+
+    if (this.containsOnlyConsiderableCharSequences(splittedStr))
+      return splittedStr;
+
+    const revelantSubSequences: string[] = [];
+
+    /* If the current subsequence is relevant, push it to revelants table
+     * Otherwise append the current one to the last relevant subsequence
+     */
+
+    splittedStr.forEach((subSequence) => {
+      if (this.isConsiderableCharSequence(subSequence))
+        revelantSubSequences.push(subSequence);
+      else
+        revelantSubSequences[revelantSubSequences.length - 1] +=
+          `${subSequence}`;
+    });
+
+    return revelantSubSequences;
   }
 }